Move practitioners getter to serviceHook file
